Add EmployeeStatusBreakdownReport model with headcount and resignation endpoints (#83)

+++ libs/drf/serializers/reports.py
@@ -4,4 +4,4 @@ from rest_framework import serializers
 class BaseStatisticsSerializer(serializers.Serializer):
     type = serializers.CharField()
     name = serializers.CharField()
-    statistics = serializers.ListField(child=serializers.IntegerField())
+    statistics = serializers.ListField(child=serializers.FloatField())
